requirements.txt
# Azure OpenAI and Document Intelligence
azure-ai-formrecognizer==3.3.2
azure-cognitiveservices-vision-computervision==0.9.0
azure-identity==1.15.0
azure-core==1.29.5
openai==0.28.1
python-dotenv==1.0.0

# Web Framework
streamlit==1.31.0
fastapi==0.109.0
uvicorn==0.27.0

# Data Processing
pandas==2.1.4
numpy==1.26.4
Pillow==10.2.0
PyPDF2==3.0.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
requests==2.31.0
python-multipart==0.0.6

# HTML Parsing
beautifulsoup4==4.12.2

# Document Processing Fallback
markitdown==0.0.1a2

# Development and Testing
pytest==7.4.4
pytest-asyncio==0.23.3